refactor: VeritalList의 ProductList 컴포넌트 모듈화
- HrizontalList 컴포넌트와의 일관성을 주고자 함
