+++ client/src/components/VerticalList/VerticalList.tsx
@@ -3,9 +3,9 @@ import { ProductCardType } from '../../types/productCard'
 import { StyledWrapper } from '../../styles/StyledWrapper'
 import { STYLES } from '../../utils/styleConstants'
 import styled from 'styled-components'
-import { ProductCard } from '../ProductCard'
+import { ProductCardList } from './ProductCardList'
 
-type Props = {
+export type Props = {
   title: string
   productList: ProductCardType[]
 }
@@ -20,13 +20,6 @@ const StyledHeader = styled.div`
     line-height: 30px;
   }
 `
-const StyledProductListWrap = styled.div`
-  justify-content: space-between;
-  display: flex;
-  flex-wrap: wrap;
-  margin-top: 10px;
-  -webkit-overflow-scrolling: touch;
-`
 
 export const VerticalList = (props: Props) => {
   const { title, productList } = props
@@ -35,18 +28,9 @@ export const VerticalList = (props: Props) => {
     <StyledWrapper>
       <StyledContainer>
         <StyledHeader>
-          <h2>{title}</h2>
+          <h2 className="title">{title}</h2>
         </StyledHeader>
-        <StyledProductListWrap>
-          {productList.map((product, idx) => (
-            <ProductCard
-              key={idx}
-              product={product}
-              width="calc(50% - 5px)"
-              style={{ marginBottom: '10px' }}
-            />
-          ))}
-        </StyledProductListWrap>
+        <ProductCardList productList={productList} />
       </StyledContainer>
     </StyledWrapper>
   )
